+++ packages/react-magma-dom/src/components/Tabs/Tab.tsx
@@ -11,10 +11,20 @@ import { ThemeInterface } from '../../theme/magma';
 
 export interface TabProps
   extends React.ButtonHTMLAttributes<HTMLButtonElement> {
+  /**
+   * Icon to display within the component
+   */
   icon?: React.ReactElement<any> | React.ReactElement<any>[];
+  /**
+   * If true, the component will display with the active/selected state
+   * @default false
+   */
   isActive?: boolean;
   isInverse?: boolean;
   testId?: string;
+  /**
+   * @internal
+   */
   theme?: any;
 }
 
@@ -229,6 +239,12 @@ export const Tab = React.forwardRef<HTMLButtonElement, TabProps>(
       isFullWidth,
     } = React.useContext(TabsContext);
 
+    const handleClick = (e, index) => {
+      changeHandler(e, index);
+
+      props.onClick && typeof props.onClick === 'function' && props.onClick(e);
+    };
+
     React.useEffect(() => {
       registerTabButton(buttonRefArray, ownRef);
 
@@ -264,7 +280,7 @@ export const Tab = React.forwardRef<HTMLButtonElement, TabProps>(
           isActive={isActive}
           isInverse={isInverse}
           isFullWidth={isFullWidth}
-          onClick={e => changeHandler(index, e)}
+          onClick={e => handleClick(index, e)}
           orientation={orientation}
           ref={ref}
           role="tab"